+++ src/demo.spec.ts
@@ -1,5 +1,0 @@
-describe("Demo", () => {
-    it("does something", () => {
-        expect(true).toBe(true);
-    });
-});