+++ CHANGELOG.md
@@ -6,4 +6,5 @@ Releases ordered so that the most recent are displayed at the top, with the curr
 
 ### Features and Improvements
 
+-   [#1](https://github.com/dbtedman/respite-client/issues/1) Requesting Respite Component
 -   [#4](https://github.com/dbtedman/respite-client/issues/4) Local Development Environment